Finalne rozwiązanie zadania domowego z dnia 15:)

+++ ConsoleApp2/ConsoleApp2/Employee.cs
@@ -44,7 +44,7 @@
         }
         public void AddScore(char score)
         {
-            switch (score)
+            switch ((float)score)
             {
                 case 'A':
                 case 'a':
